Rename page to pages
