+++ src/Expensify.js
@@ -2,8 +2,8 @@ import React, {Component} from 'react';
 import {View} from 'react-native';
 import PropTypes from 'prop-types';
 import {recordCurrentlyViewedReportID, recordCurrentRoute} from './libs/actions/App';
-import SignInPage from './page/SignInPage';
-import HomePage from './page/home/HomePage';
+import SignInPage from './pages/SignInPage';
+import HomePage from './pages/home/HomePage';
 import Ion from './libs/Ion';
 import * as ActiveClientManager from './libs/ActiveClientManager';
 import IONKEYS from './IONKEYS';
